Cocoboot: Integrated treo code (removed all ifdefs). Closes: 5
This commit adjusts the boot procedure to disable all memory protection.
This means that we do not need to make a writable copy of translation table
on treo devices anymore.  All treo #ifdefs have been removed and the
translation table base setting has been moved to mach.c as a runtime check.


git-svn-id: https://hackndev.svn.sourceforge.net/svnroot/hackndev/cocoboot/trunk@1136 4cf8e309-b20d-0410-9cb3-b925f7cd04b4