Add stub for loading ELF binaries
Linux is unaffected by this since the resulting kernel (that blob that's
actually loaded) isn't in ELF format and what's better, the decompressor
itself handles all the dirty work (relocation...).

Some operating systems have the resulting kernel as an ELF binary though
so that's why an ELF loader would be nice to have (to load them).